src/requirements.txt
Django==4.0.4
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.27.1
